services/evaluation/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
scikit-learn==1.4.2
numpy==1.26.4
psycopg2-binary==2.9.9
pydantic==2.7.0
